backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]>=2.5.0
google-generativeai==0.3.0
python-dotenv==1.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.3
passlib[bcrypt]>=1.7.4
